simplify schema + add competence to schema + move printGridDocx function
